@@ -2,7 +2,6 @@
 package org.example.shared.security;
 
 import java.time.Duration;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -11,10 +10,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
@@ -28,8 +27,9 @@ import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 @EnableMethodSecurity
 public class SecurityConfig {
 
-    @Value("${app.cors.allowed-origins:*}") // ENV: APP_CORS_ALLOWED_ORIGINS
-    private String allowedOrigins;
+    // Accept both env keys. Use APP_CORS_ALLOWED_ORIGINS first, else CORS_ALLOWED_ORIGINS, else "*".
+    @Value("${app.cors.allowed-origins:${APP_CORS_ALLOWED_ORIGINS:${CORS_ALLOWED_ORIGINS:*}}}")
+    private String allowedOriginsCsv;
 
     private final JwtAuthenticationFilter jwtAuthenticationFilter;
 
@@ -51,6 +51,8 @@ public class SecurityConfig {
                     "/v3/api-docs", "/v3/api-docs/**", "/v3/api-docs.yaml",
                     "/favicon.ico", "/webjars/**", "/actuator/**"
                 ).permitAll()
+                // Realtime token endpoints should be public if they use their own auth inside
+                .requestMatchers("/realtime/**").authenticated()
                 .anyRequest().authenticated()
             )
             .httpBasic(b -> b.disable())
@@ -63,37 +65,30 @@ public class SecurityConfig {
 
     @Bean
     CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration c = new CorsConfiguration();
+        final var cfg = new CorsConfiguration();
 
-        List<String> origins = parseOrigins(allowedOrigins);
-        if (origins.size() == 1 && "*".equals(origins.get(0))) {
-            c.addAllowedOriginPattern("*"); // dev only; avoid in prod with credentials
+        final var trimmed = Arrays.stream(allowedOriginsCsv.split(","))
+                .map(String::trim)
+                .filter(s -> !s.isEmpty())
+                .map(s -> s.endsWith("/") ? s.substring(0, s.length() - 1) : s)
+                .toList();
+
+        // Credentials require explicit origins in prod. Fall back to pattern for dev.
+        if (trimmed.size() == 1 && "*".equals(trimmed.get(0))) {
+            cfg.setAllowedOriginPatterns(List.of("*")); // dev only
         } else {
-            origins.forEach(c::addAllowedOrigin);
+            cfg.setAllowedOrigins(trimmed);
         }
 
-        c.setAllowCredentials(true);
-        c.setAllowedMethods(Arrays.asList("GET","POST","PUT","PATCH","DELETE","OPTIONS"));
-        c.setAllowedHeaders(Arrays.asList("Authorization","Content-Type"));
-        c.setExposedHeaders(Arrays.asList("ETag","Location","Link"));
-        c.setMaxAge(Duration.ofHours(1));
+        cfg.setAllowCredentials(true);
+        cfg.setAllowedMethods(List.of("GET","POST","PUT","PATCH","DELETE","OPTIONS"));
+        cfg.setAllowedHeaders(List.of("*"));              // allow all incoming request headers
+        cfg.setExposedHeaders(List.of("ETag","Location","Link","Content-Disposition"));
+        cfg.setMaxAge(Duration.ofHours(1));
 
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", c);
+        var source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", cfg);
         return source;
-    }
-
-    private List<String> parseOrigins(String csv) {
-        String[] parts = csv == null ? new String[0] : csv.split(",");
-        List<String> out = new ArrayList<>(parts.length);
-        for (String raw : parts) {
-            if (raw == null) continue;
-            String o = raw.trim();
-            if (o.endsWith("/")) o = o.substring(0, o.length() - 1);
-            if (!o.isEmpty()) out.add(o);
-        }
-        if (out.isEmpty()) out.add("*");
-        return out;
     }
 
     @Bean
